Trading_bot/_archive/task3/local_version/requirements.txt
psycopg2
asyncio>=3.0.0
python-telegram-bot>10.0
urllib3>=2.0.0
jsonref
requests
datetime
beautifulsoup4
numpy
pybit
ccxt
